requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitsandbytes==0.45.5
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.2
cycler==0.12.1
einops==0.6.1
einops-exts==0.0.4
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.13.1
fonttools==4.58.0
fsspec==2024.6.1
ftfy==6.3.1
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.31.4
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
latex2mathml==3.78.0
# Editable install with no version control (llava==1.2.2.post1)
-e /home/user/Tech/LLM/LLaVA
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.40.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
peft==0.15.2
pillow==10.4.0
poppler-utils==0.1.0
protobuf==6.31.0
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.0
ruff==0.11.10
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.6.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.37.2
triton==2.1.0
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
